# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Adare General Hospital Platform API",
    "version": "1.0.0",
    "description": "REST API for the Adare General Hospital digital platform (Hawassa, Sidama Region, Ethiopia). All responses use the envelope { ok, data, message } or { ok:false, code, error }. Protected endpoints require Authorization: Bearer <access_token>; refresh tokens rotate via HTTP-only cookie on /api/auth."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
...
  },
  "paths": {
...
    "/services": { "get": { "summary": "Searchable services directory", "parameters": [{ "name": "q", "in": "query" }, { "name": "department", "in": "query" }], "responses": { "200": { "description": "Services" } } } },
    "/doctors": { "get": { "summary": "Doctor directory", "parameters": [{ "name": "q", "in": "query" }, { "name": "department", "in": "query" }], "responses": { "200": { "description": "Doctors" } } } },
...
    "/news": { "get": { "summary": "Published news (paginated)", "parameters": [{ "name": "page", "in": "query" }, { "name": "category", "in": "query" }, { "name": "q", "in": "query" }], "responses": { "200": { "description": "Articles" } } } },
    "/news/{slug}": { "get": { "summary": "Single article", "parameters": [{ "name": "slug", "in": "path", "required": true }], "responses": { "200": { "description": "Article" }, "404": { "description": "Not found" } } } },
...
    "/search": { "get": { "summary": "Global public search", "parameters": [{ "name": "q", "in": "query", "required": true }], "responses": { "200": { "description": "Mixed results" } } } },
...
  }
}